rm all deps to clients test classpath
