+++ clients/src/testFixtures/java/org/apache/kafka/clients/MockClient.java
