+++ clients/src/testFixtures/java/org/apache/kafka/clients/consumer/internals/MockRebalanceListener.java
