+++ clients/src/testFixtures/java/org/apache/kafka/common/config/provider/MockFileConfigProvider.java
