+++ clients/src/testFixtures/java/org/apache/kafka/common/metrics/FakeMetricsReporter.java
