+++ clients/src/testFixtures/java/org/apache/kafka/common/network/CertStores.java
