+++ clients/src/testFixtures/java/org/apache/kafka/common/record/internal/ArbitraryMemoryRecords.java
