+++ clients/src/testFixtures/java/org/apache/kafka/common/record/internal/InvalidMemoryRecordsProvider.java
