+++ clients/src/testFixtures/java/org/apache/kafka/common/security/authenticator/TestJaasConfig.java
