+++ clients/src/testFixtures/java/org/apache/kafka/common/utils/LogCaptureAppender.java
