+++ clients/src/testFixtures/java/org/apache/kafka/test/MockConsumerInterceptor.java
