+++ clients/src/testFixtures/java/org/apache/kafka/test/MockDeserializer.java
