+++ clients/src/testFixtures/java/org/apache/kafka/test/MockMetricsReporter.java
