+++ clients/src/testFixtures/java/org/apache/kafka/test/MockSerializer.java
