+++ core/src/test/scala/integration/kafka/api/PlaintextAdminIntegrationTest.scala
@@ -1862,8 +1862,8 @@ class PlaintextAdminIntegrationTest extends BaseAdminIntegrationTest {
     val config = createConfig
     config.put(AdminClientConfig.DEFAULT_API_TIMEOUT_MS_CONFIG, "100000000")
     config.put(AdminClientConfig.RETRIES_CONFIG, "0")
-    val factory = new KafkaAdminClientTest.FailureInjectingTimeoutProcessorFactory()
-    client = KafkaAdminClientTest.createInternal(new AdminClientConfig(config), factory)
+    val factory = new FailureInjectingTimeoutProcessorFactory()
+    client = KafkaAdminClientInternalFactory.createInternal(new AdminClientConfig(config), factory)
     val future = client.createTopics(Seq("mytopic", "mytopic2").map(new NewTopic(_, 1, 1.toShort)).asJava,
         new CreateTopicsOptions().validateOnly(true)).all()
     assertFutureThrows(classOf[TimeoutException], future)
